Guest, Booking services validations with exceptions

+++ Hotels.Service/Implementations/BookingService.cs
@@ -16,17 +16,20 @@ namespace Hotels.Service.Implementations
         private readonly IBookingRepository _bookingRepository;
         private readonly IReservationService _reservationService;
         private readonly IRoomService _roomService;
+        private readonly IGuestService _guestService;
         private readonly IMapper _mapper;
 
         public BookingService(IBookingRepository bookingRepository, 
             IMapper mapper, 
             IReservationService reservationService, 
-            IRoomService roomService)
+            IRoomService roomService,
+            IGuestService guestService)
         {
             _bookingRepository = bookingRepository;
             _mapper = mapper;
             _reservationService = reservationService;
             _roomService= roomService;
+            _guestService = guestService;
         }
         public async Task AddBooking(BookingWithReservationAddingDto bookingWithReservationDto)
         {
@@ -36,6 +39,13 @@ namespace Hotels.Service.Implementations
             {
                 throw new RoomUnavailableException($"Room with id {bookingWithReservationDto.RoomId} is not available to reserve");
             }
+            if(roomToReserve is null)
+            {
+                throw new NotFoundException("You are chosing wrong room.");
+            }
+
+
+            
 
 
             var existingReservationsOfRoom = await _reservationService.GetReservationsOfRoom(bookingWithReservationDto.RoomId);
@@ -76,22 +86,30 @@ namespace Hotels.Service.Implementations
 
         public async Task DeleteBooking(int id)
         {
+            if (id <= 0)
+            {
+                throw new ArgumentOutOfRangeException("id can not be negative or zero");
+            }
             var bookingToDelete = await _bookingRepository.GetAsync(x => x.Id == id);
+            if(bookingToDelete is null)
+            {
+                throw new NotFoundException($"Booking with {id} was not found");
+            }
 
             int reservationIdToDelete= bookingToDelete.ReservationId;
             await _reservationService.DeleteReservation(reservationIdToDelete);
 
             _bookingRepository.Remove(bookingToDelete);
 
-
-
-
-
         }
 
         public async Task<List<BookingGettingDto>> GetAllBookings()
         {
             List<Booking> bookings=await _bookingRepository.GetAllAsync();
+            if(bookings is null)
+            {
+                throw new NotFoundException("There are not any reservations in database");
+            }
             var obj=_mapper.Map<List<BookingGettingDto>>(bookings);
             return obj;
         }
@@ -103,15 +121,25 @@ namespace Hotels.Service.Implementations
 
         public async Task<List<BookingGettingDto>> GetBookingsOfUser(int userId)
         {
+            var userToFind=await _guestService.GetGuest(userId);
             List<Booking> result = await _bookingRepository.GetAllAsync(x=>x.GuestId==userId);
+            if(result is null || result.Count==0)
+            {
+                throw new NotFoundException("Bookings for this user was not found");
+            }
             var obj = _mapper.Map<List<BookingGettingDto>>(result);
             return obj;
         }
 
         public async Task<List<ReservationGettingDto>> GetReservationsOfGuest(int guestId)
         {
+            var userToFind = await _guestService.GetGuest(guestId);
             //bookingebidan avige is idebi, romelic am useris aris
             var reservationIds = await _bookingRepository.GetBookingIdsByGuestIdAsync(guestId);
+            if(reservationIds is null  || reservationIds.Count == 0)
+            {
+                throw new NotFoundException("Reservations for this user was not found");
+            }
 
             //avige is reservaciebi, romelta aidebic aris zeda listshi
             List<ReservationGettingDto> result = new();
